+++ src/ftxui/component/maybe.cpp
@@ -9,47 +9,75 @@
 
 namespace ftxui {
 
-Component Maybe(Component child, std::function<bool()>&& do_show)
-{
+Component Maybe(Component child, std::function<bool()> show) {
   class Impl : public ComponentBase {
    public:
-    Impl(std::function<bool()>&& do_show) : do_show_(std::move(do_show)) {}
+    Impl(std::function<bool()> show) : show_(std::move(show)) {}
 
    private:
     Element Render() override {
-      return do_show_() ? ComponentBase::Render() : std::make_unique<Node>();
+      return show_() ? ComponentBase::Render() : std::make_unique<Node>();
     }
     bool Focusable() const override {
-      return do_show_() && ComponentBase::Focusable();
+      return show_() && ComponentBase::Focusable();
     }
     bool OnEvent(Event event) override {
-      return do_show_() && ComponentBase::OnEvent(event);
+      return show_() && ComponentBase::OnEvent(event);
     }
 
-    std::function<bool()> do_show_;
+    std::function<bool()> show_;
   };
 
-  auto maybe = Make<Impl>(std::move(do_show));
+  auto maybe = Make<Impl>(std::move(show));
   maybe->Add(std::move(child));
   return maybe;
 }
 
-ComponentDecorator Maybe(std::function<bool()>&& do_show) {
-  return [do_show = std::move(do_show)](Component child) mutable {
-    return Maybe(child, std::move(do_show));
+/// @brief Decorate a component. It is shown only when the |show| function
+/// returns true.
+/// @params show a function returning whether the decoratorated component should
+/// be shown.
+/// @ingroup component
+///
+/// ### Example
+///
+/// ```cpp
+/// auto component = Renderer([]{ return "Hello World!"; });
+/// auto maybe_component = component | Maybe([&]{ return counter == 42; });
+/// ```
+ComponentDecorator Maybe(std::function<bool()> show) {
+  return [show = std::move(show)](Component child) mutable {
+    return Maybe(child, std::move(show));
   };
 }
 
+/// @brief Decorate a component |child|. It is shown only when |show| is true.
+/// @params child the compoennt to decorate.
+/// @params show a boolean. |child| is shown when |show| is true.
+/// @ingroup component
+///
+/// ### Example
+///
+/// ```cpp
+/// auto component = Renderer([]{ return "Hello World!"; });
+/// auto maybe_component = Maybe(component, &show);
+/// ```
 Component Maybe(Component child, const bool* show) {
-  return Maybe(child, [show] {
-    return *show;
-  });
+  return Maybe(child, [show] { return *show; });
 }
 
+/// @brief Decorate a component. It is shown only when |show| is true.
+/// @params show a boolean. |child| is shown when |show| is true.
+/// @ingroup component
+///
+/// ### Example
+///
+/// ```cpp
+/// auto component = Renderer([]{ return "Hello World!"; });
+/// auto maybe_component = component | Maybe(&show);
+/// ```
 ComponentDecorator Maybe(const bool* show) {
-  return [show](Component child) {
-    return Maybe(child, show);
-  };
+  return [show](Component child) { return Maybe(child, show); };
 }
 
 }  // namespace ftxui